Engineering: Support clang 10 on Linux

+++ KlayGE/Core/Src/Render/ElementFormat.cpp
@@ -341,7 +341,7 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				uint32_t const s = *reinterpret_cast<uint32_t const *>(p);
-				*output = Color(static_cast<float>(s) / 0xFFFFFFFF, 0, 0, 1);
+				*output = Color(static_cast<float>(s) / static_cast<float>(0xFFFFFFFF), 0, 0, 1);
 			}
 			break;
 
@@ -349,7 +349,7 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				int32_t const s = *reinterpret_cast<int32_t const *>(p);
-				*output = Color(static_cast<float>(s) / 0x7FFFFFFF, 0, 0, 1);
+				*output = Color(static_cast<float>(s) / static_cast<float>(0x7FFFFFFF), 0, 0, 1);
 			}
 			break;
 
@@ -357,7 +357,8 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				uint32_t const * s = reinterpret_cast<uint32_t const *>(p);
-				*output = Color(static_cast<float>(s[0]) / 0xFFFFFFFF, static_cast<float>(s[1]) / 0xFFFFFFFF, 0, 1);
+				*output = Color(static_cast<float>(s[0]) / static_cast<float>(0xFFFFFFFF),
+					static_cast<float>(s[1]) / static_cast<float>(0xFFFFFFFF), 0, 1);
 			}
 			break;
 
@@ -365,7 +366,8 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				int32_t const * s = reinterpret_cast<int32_t const *>(p);
-				*output = Color(static_cast<float>(s[0]) / 0x7FFFFFFF, static_cast<float>(s[1]) / 0x7FFFFFFF, 0, 1);
+				*output = Color(static_cast<float>(s[0]) / static_cast<float>(0x7FFFFFFF),
+					static_cast<float>(s[1]) / static_cast<float>(0x7FFFFFFF), 0, 1);
 			}
 			break;
 
@@ -373,8 +375,9 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				uint32_t const * s = reinterpret_cast<uint32_t const *>(p);
-				*output = Color(static_cast<float>(s[0]) / 0xFFFFFFFF, static_cast<float>(s[1]) / 0xFFFFFFFF,
-					static_cast<float>(s[2]) / 0xFFFFFFFF, 1);
+				*output = Color(static_cast<float>(s[0]) / static_cast<float>(0xFFFFFFFF),
+					static_cast<float>(s[1]) / static_cast<float>(0xFFFFFFFF), static_cast<float>(s[2]) / static_cast<float>(0xFFFFFFFF),
+					1);
 			}
 			break;
 
@@ -382,8 +385,9 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				int32_t const * s = reinterpret_cast<int32_t const *>(p);
-				*output = Color(static_cast<float>(s[0]) / 0x7FFFFFFF, static_cast<float>(s[1]) / 0x7FFFFFFF,
-					static_cast<float>(s[2]) / 0x7FFFFFFF, 1);
+				*output = Color(static_cast<float>(s[0]) / static_cast<float>(0x7FFFFFFF),
+					static_cast<float>(s[1]) / static_cast<float>(0x7FFFFFFF), static_cast<float>(s[2]) / static_cast<float>(0x7FFFFFFF),
+					1);
 			}
 			break;
 
@@ -391,8 +395,9 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				uint32_t const * s = reinterpret_cast<uint32_t const *>(p);
-				*output = Color(static_cast<float>(s[0]) / 0xFFFFFFFF, static_cast<float>(s[1]) / 0xFFFFFFFF,
-					static_cast<float>(s[2]) / 0xFFFFFFFF, static_cast<float>(s[3]) / 0xFFFFFFFF);
+				*output = Color(static_cast<float>(s[0]) / static_cast<float>(0xFFFFFFFF),
+					static_cast<float>(s[1]) / static_cast<float>(0xFFFFFFFF), static_cast<float>(s[2]) / static_cast<float>(0xFFFFFFFF),
+					static_cast<float>(s[3]) / static_cast<float>(0xFFFFFFFF));
 			}
 			break;
 
@@ -400,8 +405,9 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, p += elem_size, ++ output)
 			{
 				int32_t const * s = reinterpret_cast<int32_t const *>(p);
-				*output = Color(static_cast<float>(s[0]) / 0x7FFFFFFF, static_cast<float>(s[1]) / 0x7FFFFFFF,
-					static_cast<float>(s[2]) / 0x7FFFFFFF, static_cast<float>(s[3]) / 0x7FFFFFFF);
+				*output = Color(static_cast<float>(s[0]) / static_cast<float>(0x7FFFFFFF),
+					static_cast<float>(s[1]) / static_cast<float>(0x7FFFFFFF), static_cast<float>(s[2]) / static_cast<float>(0x7FFFFFFF),
+					static_cast<float>(s[3]) / static_cast<float>(0x7FFFFFFF));
 			}
 			break;
 
@@ -1065,7 +1071,8 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				uint32_t* s = reinterpret_cast<uint32_t*>(p);
-				*s = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->r() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				*s = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->r() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
 			}
 			break;
 
@@ -1073,7 +1080,8 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				int32_t* s = reinterpret_cast<int32_t*>(p);
-				*s = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->r() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				*s = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->r() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
 			}
 			break;
 
@@ -1081,8 +1089,10 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				uint32_t* s = reinterpret_cast<uint32_t*>(p);
-				s[0] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->r() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
-				s[1] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->g() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[0] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->r() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[1] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->g() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
 			}
 			break;
 
@@ -1090,8 +1100,10 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				int32_t* s = reinterpret_cast<int32_t*>(p);
-				s[0] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->r() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
-				s[1] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->g() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[0] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->r() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[1] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->g() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
 			}
 			break;
 
@@ -1099,9 +1111,12 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				uint32_t* s = reinterpret_cast<uint32_t*>(p);
-				s[0] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->r() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
-				s[1] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->g() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
-				s[2] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->b() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[0] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->r() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[1] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->g() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[2] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->b() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
 			}
 			break;
 
@@ -1109,9 +1124,12 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				int32_t* s = reinterpret_cast<int32_t*>(p);
-				s[0] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->r() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
-				s[1] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->g() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
-				s[2] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->b() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[0] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->r() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[1] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->g() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[2] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->b() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
 			}
 			break;
 
@@ -1119,10 +1137,14 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				uint32_t* s = reinterpret_cast<uint32_t*>(p);
-				s[0] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->r() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
-				s[1] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->g() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
-				s[2] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->b() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
-				s[3] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(static_cast<uint64_t>(input->a() * 0xFFFFFFFFULL + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[0] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->r() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[1] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->g() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[2] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->b() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
+				s[3] = static_cast<uint32_t>(MathLib::clamp<uint64_t>(
+					static_cast<uint64_t>(input->a() * static_cast<float>(0xFFFFFFFFULL) + 0.5f), 0ULL, 0xFFFFFFFFULL));
 			}
 			break;
 
@@ -1130,10 +1152,14 @@ namespace KlayGE
 			for (uint32_t i = 0; i < num_elems; ++ i, ++ input, p += elem_size)
 			{
 				int32_t* s = reinterpret_cast<int32_t*>(p);
-				s[0] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->r() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
-				s[1] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->g() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
-				s[2] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->b() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
-				s[3] = static_cast<int32_t>(MathLib::clamp<int64_t>(static_cast<int64_t>(input->a() * 0x7FFFFFFFLL + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[0] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->r() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[1] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->g() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[2] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->b() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
+				s[3] = static_cast<int32_t>(MathLib::clamp<int64_t>(
+					static_cast<int64_t>(input->a() * static_cast<float>(0x7FFFFFFFLL) + 0.5f), -0x7FFFFFFFLL, 0x7FFFFFFFLL));
 			}
 			break;
